+++ src/simmate/website/core/models/__init__.py
@@ -3,4 +3,5 @@
 from simmate.database.mixins import Fingerprint, FingerprintPool, Spacegroup
 
 from .api_token import ApiToken
+from .notification import Notification
 from .website_page_visits import WebsitePageVisit